Move fillArrays function into a separate file
